+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- In web UI, on event detail page is said "Edits Approved" when it should have said "Edits Refused"
+
 ### Changed
 
 - In web UI, order records